refactor: Move and rename files
Movido dois arquivos para novos diretórios e atualizado o nome das variáveis para melhor refletir suas funções e contextos.
